refactor: move server handler tree to internal/server (#51)
Mechanical relocation of pkg/goSignals/server to internal/server,
restoring the pkg/* must-not-import-internal/* invariant from CLAUDE.md.
pkg/goSignals/ now contains only api/, swagger-ui/, swagger.go,
and the Dockerfile.

No behavior changes - same exports, same HTTP wire surface, same vet
warnings. Updates 4 external import sites plus 4 inside the moved
test/ subpackage. tls_logging_test.go relative cert paths shorten by
one ../ level to match the new depth.

CONTEXT.md gains a Package boundaries section noting the new location
and the pkg/goSsfServer -> internal/server cross-boundary import as
known debt (PRD #50 Out of Scope).

Part of PRD #50.
