+++ internal/server/test/load_key_extension_test.go
