+++ internal/server/test/operational_event_test.go
