+++ internal/server/test/poll_discovery_test.go
