+++ internal/server/test/poll_infinite_loop_fix_test.go
