+++ internal/server/test/register_client_test.go
