+++ internal/server/test/resources/mongo_token.json
