+++ internal/server/test/stream_update_external_token_test.go
