+++ internal/server/test/trigger_event_test.go
