+++ pkg/goSsfServer/ssf-application.go
@@ -17,7 +17,7 @@ import (
 	"github.com/i2-open/i2goSignals/internal/providers/storage"
 	"github.com/i2-open/i2goSignals/internal/services"
 	"github.com/i2-open/i2goSignals/pkg/constants"
-	"github.com/i2-open/i2goSignals/pkg/goSignals/server"
+	"github.com/i2-open/i2goSignals/internal/server"
 	"github.com/i2-open/i2goSignals/pkg/logger"
 	"github.com/i2-open/i2goSignals/pkg/ssfModels"
 )
